Update build script to create internally testable 'upm' packages
Summary: Update the build script to generate debug *.tgz packages for use with the new installer. This also cleans up the package names to be something more external dependent. We'll need to follow up and add a way to get UPM packages for a release build.

Reviewed By: ryanbailey-fb

Differential Revision:
D44775961

Privacy Context Container: L1079626

fbshipit-source-id: 259131a498c463b4dc3c61dd4705205cef786167