Adding package descriptions
Summary:
Adding descriptions for the packages as that's useful for UPM.  :)
For all the public ones, I simply copied the descriptions from the website.
There are also descriptions added to the various libs which are not intending to be published publicly but are instead rolled into the higher level packages -- just in case.

These will need to be picked back into v55/v56

Reviewed By: rokeyge

Differential Revision: D47258434

fbshipit-source-id: a3c7f219a4e07929527994872de174b514aaa1b6